Table structure changes, got authorization codes working, I think

+++ Spotifly/Models/User.cs
@@ -10,7 +10,7 @@ namespace Spotifly.Models
 {
     public class User
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
@@ -27,7 +27,7 @@ namespace Spotifly.Models
             }
             return users;
         }
-
+        
         public static List<User> SelectFriends(string id)
         {
             string sql = "SELECT id, name, date_created FROM user WHERE user1_id == @UserId OR user2_id == @UserId";
@@ -39,7 +39,6 @@ namespace Spotifly.Models
             }
             return users;
         }
-
         public static User SelectByEmailForAuth(string email)
         {
             string sql = "SELECT id, name, email, password, date_created FROM user WHERE email = @Email";
@@ -51,7 +50,6 @@ namespace Spotifly.Models
             }
             return user;
         }
-
         public static User SelectByEmail(string email)
         {
             string sql = "SELECT id, name, email, date_created FROM user WHERE email = @Email";
@@ -79,7 +77,7 @@ namespace Spotifly.Models
             return user;
         }
 
-        public static bool UpdatePassword(int id, string password)
+        public static bool UpdatePassword(Int64 id, string password)
         {
             string sql = "UPDATE user SET password=@Password where id=@Id";
             bool success = false;
@@ -103,7 +101,7 @@ namespace Spotifly.Models
             }
             return success;
         }
-        public static bool UpdateEmail(int id, string email)
+        public static bool UpdateEmail(Int64 id, string email)
         {
             string sql = "UPDATE user SET email=@Email where id=@Id";
             bool success = false;
